Add ec2debian-build-ami plugin
Using ec2debian-build-ami[1] it is possible to build an Amazon Machine Image
automatically running the Warrior on startup. Since both the Warrior and the
AMI are Debian based and use almost the same setup, everything that runs on the
Warrior VM will also run on the Warrior AMI, including precompiled binaries.

Creating a new AMI:
    sudo ./ec2debian-build-ami --arch i386 --plugin plugins/admin-user \
    --plugin ../ec2debian-build-ami-plugin/warrior-setup --name warrior \
    --access-key XXX -secret-key XXX
Example userdata:
    {"downloader": "ec2-warrior", "selected_project": "auto",
    "concurrent_items": "6", "shared:rsync_threads": "4"}

1: https://github.com/andsens/ec2debian-build-ami